Move setting policy action to application code
